+++ src/checker/base/__init__.py
@@ -1,3 +1,3 @@
-from .base_checker import BaseChecker, BaseLevel2Checker, BaseLevel3Checker
+from .checker import BaseChecker, BaseLevel2Checker, BaseLevel3Checker
 
 __all__ = ["BaseChecker", "BaseLevel2Checker", "BaseLevel3Checker"]
