+++ src/checker/common.py
