--- src/checker/csv/__init__.py
@@ -1,3 +1,0 @@
-# Level1、Level2、Level3チェッカーは全て統合されました
-
-__all__ = []
