--- src/checker/xls/__init__.py
@@ -1,4 +1,0 @@
-# XLS専用チェッカーのエントリポイント
-# Level1、Level2、Level3チェッカーは全て統合されました
-
-__all__ = []
