--- src/checker/xlsx/__init__.py
@@ -1,4 +1,0 @@
-# XLSX専用チェッカーのエントリポイント
-# Level1、Level2、Level3チェッカーは全て統合されました
-
-__all__ = []
